fix(s99): s99_flag1 bitfields must use unsigned short not int
Under #pragma pack(1) xlc allocates a 4-byte container for int
bitfields, making s99_flag1 occupy 4 bytes instead of the 2 bytes
IBM defines at IEFZB4D2 offset 0x05.  This misaligns s99error,
s99info, s99txtpp and s99s99x in the s99rb struct, causing SVC99 to
store the RBX pointer at the correct offset while C reads it from
+2 bytes away — producing a garbled RBX pointer whose eye-catcher
appears as 'i99RBX' instead of 'S99RBX', causing IEFDB476 to reject
the RBX with rc=0x0C on every call to s99_prt_msg.

Fix: change all s99_flag1 bitfield member types from int to
unsigned short so the container is exactly 2 bytes.

+++ core/include/s99.h
@@ -19,20 +19,23 @@ typedef enum s99_verb {
 	S99VRBIN = 7
 };
 
+/* s99_flag1 maps to a 2-byte field in the IBM SVC99 RB (IEFZB4D2 offset 0x05).
+ * Using 'unsigned short' forces a 2-byte bitfield container under pack(1);
+ * 'int' would allocate 4 bytes, misaligning s99error/s99info/s99s99x by +2. */
 struct s99_flag1 {
-	int s99oncnv:1;
-	int s99nocnv:1;
-	int s99nomnt:1;
-	int s99jbsys:1;
-	int s99cnenq:1;
-	int s99gdgnt:1;
-	int s99msglo:1;
-	int s99nomig:1;
-	int s99nosym:1;
-	int s99acucb:1;
-	int s99dsaba:1;
-	int s99dxacu:1;
-	int s99rsrv:4;
+	unsigned short s99oncnv:1;
+	unsigned short s99nocnv:1;
+	unsigned short s99nomnt:1;
+	unsigned short s99jbsys:1;
+	unsigned short s99cnenq:1;
+	unsigned short s99gdgnt:1;
+	unsigned short s99msglo:1;
+	unsigned short s99nomig:1;
+	unsigned short s99nosym:1;
+	unsigned short s99acucb:1;
+	unsigned short s99dsaba:1;
+	unsigned short s99dxacu:1;
+	unsigned short s99rsrv:4;
 };
 
 struct s99_flag2 {
